add product on shopview if admin
